Implement probability averaging

Given a set of cumulative distribution functions (assume for now: defined on same values), combine using probability averaging (also called linear opinion pool). #TODO CITE This method calculates the (weighted) average of quantiles at a given value

Usage

LOP(
  quantile,
  value,
  id,
  ret_quantiles,
  ret_values,
  weight_fn = equal_weights,
  ...
)

Arguments

quantile

vector containing quantiles for all cdfs to be aggregated

value

vector containing values for all cdfs to be aggregated

id

vector containing unique ids to distinguish cdf to be aggregated

ret_quantiles

vector of quantiles to return specifying the aggregate distribution

ret_values

vector of values to return specifying the aggregate distribution

weight_fn

function? specifying how to weight each model *FIX

...

additional arguments

Value

vector of values for corresponding ret_quantiles of aggregate distribution

Examples

dat <- expand.grid(id = c("A", "B"),
                   quantile = seq(0,1,0.01))
dat$value <- ifelse(dat$id == "A", qnorm(dat$quantile), qnorm(dat$quantile, 0,2))
LOP(dat$quantile, dat$value, dat$id, seq(0,1,0.05), NA)
#> # A tibble: 21 × 2
#>    quantile    value
#>       <dbl>    <dbl>
#>  1     0    -Inf    
#>  2     0.05  NaN    
#>  3     0.1    -1.90 
#>  4     0.15   -1.48 
#>  5     0.2    -1.17 
#>  6     0.25   -0.924
#>  7     0.3    -0.710
#>  8     0.35   -0.518
#>  9     0.4    -0.339
#> 10     0.45   -0.168
#> # ℹ 11 more rows