aggregate a set of cdfs

Given a data.frame containing cdfs , return a single aggregate cdf using specified method.

Usage

aggregate_cdfs(
  data,
  id_var,
  group_by = NULL,
  method,
  ret_quantiles,
  ret_values = NA,
  weighting_scheme = "equal",
  reorder_quantiles = TRUE,
  ...
)

Arguments

data

data.frame that contains multiple cdfs, grouped by by column. Specify cdf with quantile and value columns,

id_var

string containing the name of the column that identifies unique cdfs

group_by

vector containing the names of the columns to create unique aggregates for

method

character name of the method for aggregation (see details).

ret_quantiles

vector of quantiles to return specifying the aggregate distribution

ret_values

vector of values to return specifying the aggregate distribution

weighting_scheme

string to indicate how to weight in the aggregate (see details).

reorder_quantiles

TRUE to ensure that quantiles are ordered; set to FALSE if quantiles in data are already ordered (increasing)

...

additional arguments

Value

TBD

Details

Methods include:

"LOP" - simple probability averaging, also called Linear Opinion Pool.

"vincent" - simple quantile averaging, also called Vincent average.

Weighting schemes include:

"equal" - equal weighting of all models and values

"user_defined" - user supplies weights applied to each model (additional input weights that is a data.frame containing id and weight columns)

trimming - "cdf_interior", "cdf_exterior", "mean_interior", "mean_exterior", following REF (additional inputs ...)